requirements.txt
Django==1.7
Fabric==1.10.0
MySQL-python==1.2.5
Pillow==2.5.3
South==0.8.4
argparse==1.2.1
django-debug-toolbar==1.2.1
ecdsa==0.11
gunicorn==19.1.1
numpy==1.9.0
paramiko==1.15.1
pycrypto==2.6.1
sqlparse==0.1.12
wsgiref==0.1.2
djangorestframework==2.4.3
markdown==2.5.1
django-filter==0.8